ON-457: refactor tests structure
