+++ .github/workflows/master.yaml
@@ -15,20 +15,15 @@ jobs:
         with:
           cache: npm
           node-version: 16
-          registry-url: 'https://npm.pkg.github.com'
-          scope: '@oriumnetwork'
       - name: Install Dependencies
         run: npm ci
-        env:
-          NODE_AUTH_TOKEN: ${{ secrets.GHB_TOKEN }}
+      - name: Code Generator
+        run: npm run codegen:goerli
+      - name: Build
+        run: npm run build:goerli
+      - name: Test
+        run: npm run test
       - name: Subgraph Auth
         run: yarn graph-auth -- ${{ secrets.THEGRAPH_API_KEY }}
-        # Goerli
-      - name: Code Generator Goerli
-        run: npm run codegen-goerli
-      - name: Build Goerli
-        run: npm run build-goerli
-      - name: Goerli Test
-        run: npm run test-goerli
-      - name: Subgraph Deploy Goerli
-        run: npm run deploy-goerli
+      - name: Subgraph Deploy
+        run: npm run deploy:goerli
