+++ tests/erc7432/revoke-handler.test.ts
@@ -1,10 +1,10 @@
 import { assert, describe, test, clearStore, afterAll, beforeEach } from 'matchstick-as'
-import { generateId } from '../../../src/utils/helper'
-import { createNewRoleGrantedEvent, createNewRoleRevokedEvent } from '../../mocks/events'
-import { ZERO_ADDRESS } from '../../../src/utils/constants'
-import { handleRoleGranted, handleRoleRevoked } from '../../../src/erc7432'
+import { generateId } from '../../src/utils/helper'
+import { createNewRoleGrantedEvent, createNewRoleRevokedEvent } from '../mocks/events'
+import { ZERO_ADDRESS } from '../../src/utils/constants'
+import { handleRoleGranted, handleRoleRevoked } from '../../src/erc7432'
 import { BigInt, Bytes } from '@graphprotocol/graph-ts'
-import { createMockNft } from '../../mocks/entities'
+import { createMockNft } from '../mocks/entities'
 
 const tokenId = '123'
 const address1 = '0x1111111111111111111111111111111111111111'
